webpos/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Pymysql
Pillow
requests 
WeasyPrint
Flask-Mail 